# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/sale": {
            "post": {
                "description": "Creates a new sale with product id, quantity, unit price, payment method, and date of sale",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale"
                ],
                "summary": "Create sale",
                "parameters": [
                    {
                        "description": "Product ID, Quantity, Unit price, Payment method, Sold at",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateSaleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.CreateSaleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/v1/sale/{id}": {
            "get": {
                "description": "Returns sale by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale"
                ],
                "summary": "Get sale",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetSaleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes sale by ID",
                "tags": [
                    "sale"
                ],
                "summary": "Delete sale",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "request.CreateSaleRequest": {
            "type": "object",
            "required": [
                "payment_method",
                "product_id",
                "quantity",
                "sold_at",
                "unit_price"
            ],
            "properties": {
                "payment_method": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "sold_at": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                }
            }
        },
...
        "response.CreateSaleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "sold_at": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.GetSaleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "sold_at": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                }
            }
        },
...
    }
}